working example, now refactoring of items generation is required
